{{cookiecutter.project_slug}}/requirements/test.txt
-r base.txt

# tool for code style
flake8==3.3.0

# testing
coverage==4.4.1
factory-boy==2.8.1
pytest-cov==2.5.1
pytest-sugar==0.8.0
pytest-django==3.1.2
django-test-plus==1.0.17
